Implement initial logic for write case of cdt_memcpy.
Send write_req to manager for non-manager machines if they don't have R/W access to the machine. If the page is R/O then logic for invalidation has been implemented but not tested.

+++ include/host.h
@@ -12,8 +12,8 @@ typedef struct cdt_server_t cdt_server_t;
 #define INVALID_PAGE 0
 #define READ_ONLY_PAGE 1
 #define READ_WRITE_PAGE 2
-#define SHARED_VA_TO_IDX(va) ((va - CDT_SHARED_VA_START) / PAGESIZE)
-#define PGROUNDDOWN(a) (((a)) & ~(PAGESIZE-1))
+#define SHARED_VA_TO_IDX(va) (((uint64_t)va - CDT_SHARED_VA_START) / PAGESIZE)
+#define PGROUNDDOWN(a) ((((uint64_t)a)) & ~(PAGESIZE-1))
 
 extern const char* const cdt_task_queue_names[CDT_MAX_MACHINES];
